requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
greenlet==3.2.3
h11==0.16.0
idna==3.10
lxml==6.0.0
outcome==1.3.0.post0
packaging==25.0
playwright==1.53.0
pyee==13.0.0
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.4
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
